Move docker check to its own package
